TensorFlow_OpenVINO/requirements.txt
numpy
tqdm
tensorflow-cpu==1.15
argparse
scipy
imageio
moviepy